Example binary showing log and map being co-hosted
This is a demo application that emulates a package repository. Each
entry in the log has a {module, version, hash}, and the map simply
enumerates all indices where a given module appears.

This uses Tessera POSIX implementation for the log, and hooks up the
LogReader for that as an InputLog to the Verifiable Index.

This starts a web server that hosts a form allowing a package name to be
input, and the list of indices are returned. Paths are:

 - /vindex/ : the verifiable index
 - /log/ : the log files, served via HTTP too

This will make it easier to prototype a demo client that will call into
this log/map combo.

TODO (in a future PR):
 - Return inclusion proofs on vindex lookup
 - Implement an output log