Added provisional API definition and client
This replaces the web server that was hosted before, and starts moving
things in a direction that is closer to what we are likely to run.

The API lookup object is defined, but is largely empty for now. Future
PRs will work to fill out the remaining fields in there, and then we can
implement full verification logic in the client.

Bumped Tessera to latest version.