[VIndex] Refactoring and cleaning tech debt
The next milestone is to add the output log. While prototyping this, it became clear that a lot of the naming is confusing. Renamed to make clear what is the input log, what is the write ahead log, etc. This makes space to add the output log parameters etc afterwards.

As part of this, simplified the storage flags so that it takes only a root directory and then creates the input and output logs in there, along with the WAL file. Other map persistence will likely be added to this directory too in the future.

Use skey->vkey method from formats, and deleted local method.

Added a high level test for the verifiable index.