Move state out of the Verifiable Index
This moves things to be shaped more like a pipleine; the mapper is
responsible for reading from the log, verifying the contents, and
passing state to build a new map via clear mechanisms. At the moment,
this is just the WAL, but the raw log checkpoint will also need to be
passed through so that it can make it to the output log.

As part of this, a compact range of the currently verified state will
also need to be persisted so that we can load that to incrementally
verify progress after a restart.