[VIndex] Refactoring and cleaning tech debt (#13)
* [VIndex] Refactoring and cleaning tech debt

The next milestone is to add the output log. While prototyping this, it became clear that a lot of the naming is confusing. Renamed to make clear what is the input log, what is the write ahead log, etc. This makes space to add the output log parameters etc afterwards.

As part of this, simplified the storage flags so that it takes only a root directory and then creates the input and output logs in there, along with the WAL file. Other map persistence will likely be added to this directory too in the future.

Use skey->vkey method from formats, and deleted local method.

Added a high level test for the verifiable index.

* Move state out of the Verifiable Index

This moves things to be shaped more like a pipleine; the mapper is
responsible for reading from the log, verifying the contents, and
passing state to build a new map via clear mechanisms. At the moment,
this is just the WAL, but the raw log checkpoint will also need to be
passed through so that it can make it to the output log.

As part of this, a compact range of the currently verified state will
also need to be persisted so that we can load that to incrementally
verify progress after a restart.