Make WAL Writer easier to initialize
Changed to use a factory method

Changed validation to allow truncating any trailing junk after the last
newline. This makes it more tolerant when recovering from a partial
fsync during a previous write.

It also demonstrates that we can truncate the file, which will be useful
in the next PR. When ingesting entries from the log, we want to ensure
that they match the input log checkpoint commitment. To do this, we will
construct a compact range while performing all of the mappings. The WAL
will be written as soon as information is available, and periodically we
will write the compact range and tree size to disk. This means that on
recovery, we may need to throw away any entries in the WAL that are
beyond this compact range commitment. This is where truncation will be
needed, and this validation code will be provided with an index to
truncate everything after.