[Vindex] Added Output Log to commit to map state
The Output Log is a POSIX Tessera log in the provided demo, but can be
changed for any other implementation that matches the OutputLog
interface.

The OutputLog is updated any time the vindex is updated. The value
written to the new leaf is the concatenation of:
 1. hex encoding of the map root
 2. a couple of newline characters
 3. the raw checkpoint from the InputLog at which the vindex was built

This includes a refactoring to push the checkpoint parsing into the
Input and Output Log definitions to improve encapsulation.

This unlocks what we need to start generating fully comprehensive proofs
for Lookup operations.