Lookup returns mostly provable response (#25)
* Lookup returns mostly provable response

The client now:
 - Verifies the output log checkpoint
 - Verifies the output leaf via inclusion proof in the output log
 - Obtains the vindex root hash and hashes the vindex leaf

The missing part is to generate the inclusion proof in the vindex on
Lookup on the server, and then to verify this on the client.

Once this is done, the remaining client work is to support a flag to
actually dereference all pointers into the input log after the last seen
size. This then forms the basis of a re-runnable self-hosted monitor.

* Review comments