[VIndex client] Deference input log pointers (#26)
This allows the pointers into the input log to be automatically
dereferenced by the client tool. For now, this just prints out the
string value of whatever is there. This isn't appropriate for every
input log, but it's good enough to demo the service end-to-end, which is
its main purpose in life.