CT demo for VIndex
This doesn't work well at the moment because the size of CT logs is too
big for this.

In order to make this manageable, the MapFn is restricted to only index
domains ending in `.co.uk`, and only outputs the full key, rather than
a key for each level in the domain hierarchy.

Even with these limitations, this has value because it provides a
working base point from which to iterate.