Verifiable Index for SumDB (#44)
This one reads straight from the log (with a slight indirection via the local proxy).
This obsoletes the version that reads from the CloneDB in
trillian-examples.

The server brings up a tlog-tiles proxy for SumDB on the same port it
serves the index and output log on. This allows the index client to
read from this to dereference the leaf indices returned by the index.