[VIndex] Custom SumDB client (#47)
Added a custom client for SumDB targeted at module maintainers. This
queries a local checkout of a Go Module, and checks that every entry in
SumDB has a corresponding git tag.

It also prints out informational output in the case that there are local
tags not present in SumDB. This is not a problem for SumDB, but it
should be interesting to module authors if a version was never served.

Note: the vanilla client still works and is documented for completeness.