[VIndex] Cleaner Output Log usage (#58)
The Output Log was being updated every time the map was updated. Now it is skipped if no data has changed in the map. This uses in-memory state, so on restarting the server, a single duplicate may still be written.

This also ensures that updates to the Output Log are performed within the same lock used to update the VIndex. This resolves a race condition that could allow Lookup to return a proof to a different index root. Used this opportunity to simply pass the Input Log checkpoint into `publish` instead of reading it from disk again.